refactor: refactoring api, helper and view code
